+++ target_chains/ethereum/entropy_sdk/solidity/EntropyStructs.sol
@@ -31,16 +31,21 @@ contract EntropyStructs {
     }
 
     struct Request {
+        // Storage slot 1 //
         address provider;
         uint64 sequenceNumber;
         // The number of hashes required to verify the provider revelation.
         uint32 numHashes;
+        // Storage slot 2 //
         // The commitment is keccak256(userCommitment, providerCommitment). Storing the hash instead of both saves 20k gas by
         // eliminating 1 store.
         bytes32 commitment;
+        // Storage slot 3 //
         // If nonzero, the randomness requester wants the blockhash of this block to be incorporated into the random number.
-        // Note that we're using a uint128 such that this field fits into the same storage slot as numHashes above.
-        // Although block.number returns a uint256, 128 bits should be plenty to index all of the blocks ever generated.
-        uint128 blockNumber;
+        // Note that we're using a uint96 such that we have an additional 20 bytes of storage afterward for an address.
+        // Although block.number returns a uint256, 96 bits should be plenty to index all of the blocks ever generated.
+        uint96 blockNumber;
+
+        // TODO: store the calling contract address here and authenticate the reveal method
     }
 }
